Move tests to new tests class
